Changes to how coords are found
The program will now pick an orientation randomly and try to find coordinates that will work with that orientation. This stops all of the orientations being either a) the exact same or b) on the same place i.e. either left/right or up/down. Also added a few safety nets if the wordfile is empty or if the supplied ArrayList is empty and so on. If you spot anything wrong with the code or something that could help flag it as an issue or make a pull request. I've left the old version on github, you can find it under code>release.